requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
phonenumbers==8.13.14
Pillow==9.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
twilio==8.3.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
gunicorn==19.9.0
psycopg2-binary==2.9.1

